first commit with actual program
This is first commit with an actual program. This is mostly my files
from a resurrected project named emeta. I made a simple parser for part
of the grammar of gnu makefile. I included so examples I found in the
gnu makefile manual. Next of the list is transitioning all of the
Euphoria code to Node Javascript and CoffeeScript.